requirements.txt
numpy==1.23.5
opencv_python==4.7.0.68
Pillow==9.3.0
streamlit==1.29.0
transformers==4.37.1
